requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
interchange==2021.0.4
monotonic==1.6
packaging==24.2
pansi==2024.11.0
pillow==11.0.0
py2neo==2021.2.4
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
